+++ iosApp/iosApp/Features/Feed/FeedViewModel.swift
@@ -16,6 +16,8 @@ class FeedViewModel: ObservableObject {
     @Published var isSendNotificationShown = false
     @Published var events = [Event]()
     @Published var topics = [Topic]()
+    @Published var cotisantConfigurations = [CotisantConfiguration]()
+    @Published var ticketConfigurations = [TicketConfiguration]()
     
     func onAppear() {
         AnalyticsService.shared.log(.screenView(screenName: "feed", screenClass: "FeedView"))
@@ -34,6 +36,18 @@ class FeedViewModel: ObservableObject {
             let topics = try await CacheService.shared.apiService().getTopics(offset: 0, limit: 10)
             DispatchQueue.main.async {
                 self.topics = topics
+            }
+        }
+        Task {
+            let cotisantConfigurations = try await CacheService.shared.apiService().getCotisantConfigurations()
+            DispatchQueue.main.async {
+                self.cotisantConfigurations = cotisantConfigurations
+            }
+        }
+        Task {
+            let ticketConfigurations = try await CacheService.shared.apiService().getTicketConfigurations()
+            DispatchQueue.main.async {
+                self.ticketConfigurations = ticketConfigurations
             }
         }
     }
