+++ iosApp/iosApp/Features/RootView.swift
@@ -26,10 +26,6 @@ struct RootView: View {
                 .tabItem {
                     Label("Clubs", systemImage: "bicycle")
                 }
-            ShopView()
-                .tabItem {
-                    Label("Boutique", systemImage: "cart")
-                }
             AccountView(viewModel: AccountViewModel(
                 saveToken: viewModel.saveToken
             ))
